Add some minimal error handling to UserListComponent
Previously all that we did if there was an error in getting the users from the server was to do a `console.log()` and move on. This changes things so that we at least gives the user a simple error message.

I feel like we should throw an error here as well so that there some stack info in the browser to debug with, but I can't figure out how to test that at the moment, so I'm going to leave it out. If someone knows how to make this work that would be great.