Clean up Karma tests
The Karma tests had gotten kinda random in places, and this is an effort to clean at least some of that up. I also added/extended comments in a number of places in an effort to try to make the code more instructive/informative.

A _lot_ of the changes to UserServiceSpec are just white space and indentation. I added a bunch of `describe()`s in an attempt to better structure things, and that altered the indentation of just about everything.