Roll back to v1 of the lockfile
The update from v11 to v12 of Angular only works with v6 of npm, which apparently wants to use v1 of the lockfile format. So I'm going to go ahead and let this roll back for now and see where we end up.

I'm pretty sure we'll move forward again at a later step in the Angular update.