Update the ESLint dependencies
It turns out that I was missing some things, and other dependencies just weren't up where they needed to be. Creating a new, blank project (as suggested by @floogulinc) guided the necessary updates to get `ng lint` to run.

Comparing to the `.eslintrc.json` suggested that I probably wanted to add the `@angular-eslint/recommended` plugin as well, so this also does that.