Fix ESLint warnings
After I got ESLint/`ng lint` working, there were a few warnings that needed to be addressed.

The interesting one was that it didn't like the empty `ngOnIniti()` method in our `UserCardComponent`. Looking at it, it doesn't look like we actually made any use of the `OnInit` interface, so I just removed that and the `ngOnInit()` method.

This further supports the idea expressed elsewhere that our cards don't really "do" much, which makes things like testing them, etc., kinda odd.

I'm not sure what the "fix" is there. Are cards mostly just "dumb" views that don't "do anything"? Or would it make sense to add some functionality to the card? Can we do so without it seeming really forced?