Remove some duplicate code in `user-list.component.spec.ts`
SonarLint was fussy about some duplicated code in this spec file, so I cleaned that up.  In the process I removed `fixture` a "global" variable in the file, which is nice. I also added a pile of comments that (hopefully?) clarify what's going on here.

Technically the call to `detectChanges()` is only required in the first `describe()`. If I didn't want `fixture` to be "global", though, it wasn't easy avoid including it in `setupUserList()`, so I just left it there.