Update many files
This commit includes some rearranging of contents of files. Sometimes, I did this to address a specific warning. Other times, I rearranged to make things a bit more readable or to be consistent with the Iteration Template layout. There are some things in this commit that relate to getting rid of flex-layout.