Copy `dependabot.yml` over from iteration template
The `dependabot.yml` file was super specific to particular versions, which I think was to reduce the noise from Dependabot updates on this repo since's it's sorta a combo copy of Labs 2 and the iteration template. That said, it wasn't being updated consistently, so I just copied things over from the iteration template.

I think the _real_ solution here is bringing all the labs repos into a mono-repo, and then we would only have one version of all these dependencies to update.